move the Building and Unit macros their own file
