+++ compiler/src/macros/Building.ts
@@ -1,60 +1,9 @@
-import { camelToDashCase, discardedName } from "../utils";
-import { InstructionBase, OperationInstruction } from "../instructions";
-import { IScope, IValue, TValueInstructions } from "../types";
-import { LiteralValue, ObjectValue, StoreValue } from "../values";
+import { IScope, IValue } from "../types";
+import { LiteralValue, ObjectValue } from "../values";
 import { MacroFunction } from "./Function";
-import { operatorMap } from "../operators";
 import { CompilerError } from "../CompilerError";
 import { ValueOwner } from "../values/ValueOwner";
-
-export const itemNames = [
-  "copper",
-  "lead",
-  "metaglass",
-  "graphite",
-  "sand",
-  "coal",
-  "titanium",
-  "thorium",
-  "scrap",
-  "silicon",
-  "plastanium",
-  "phaseFabric",
-  "surgeAlloy",
-  "sporePod",
-  "blastCompound",
-  "pyratite",
-];
-
-export class Building extends ObjectValue {
-  toString() {
-    return this.owner?.name ?? discardedName;
-  }
-
-  constructor(scope: IScope) {
-    super(scope, {
-      $get: new MacroFunction(scope, (prop: IValue) => {
-        if (prop instanceof LiteralValue && typeof prop.data === "string") {
-          const name = itemNames.includes(prop.data)
-            ? camelToDashCase(prop.data)
-            : prop.data;
-          const temp = new StoreValue(scope);
-          return [
-            temp,
-            [new InstructionBase("sensor", temp, this, `@${name}`)],
-          ];
-        }
-        if (prop instanceof StoreValue) {
-          const temp = new StoreValue(scope);
-          return [temp, [new InstructionBase("sensor", temp, this, prop)]];
-        }
-        throw new CompilerError(
-          "Building property acessors must be string literals or stores"
-        );
-      }),
-    });
-  }
-}
+import { Building } from "./Entities";
 
 export class BuildingBuilder extends ObjectValue {
   constructor(scope: IScope) {
@@ -73,22 +22,4 @@ export class BuildingBuilder extends ObjectValue {
       }),
     });
   }
-}
-
-for (const key in operatorMap) {
-  type K = keyof typeof operatorMap;
-  const kind = operatorMap[key as K];
-  Building.prototype[key as K] = function (
-    this: Building,
-    scope: IScope,
-    value: IValue
-  ): TValueInstructions {
-    this.ensureOwned();
-    const [right, rightInst] = value.consume(scope);
-    const temp = new StoreValue(scope);
-    return [
-      temp,
-      [...rightInst, new OperationInstruction(kind, temp, this, right)],
-    ];
-  };
 }
