+++ compiler/src/macros/commands/UnitLocate.ts
@@ -2,8 +2,8 @@ import { InstructionBase } from "../../instructions";
 import { MacroFunction } from "..";
 import { IScope, IValue } from "../../types";
 import { LiteralValue, ObjectValue, StoreValue } from "../../values";
-import { Building } from "../Building";
 import { CompilerError } from "../../CompilerError";
+import { Building } from "../Entities";
 
 const validFinds = ["ore", "building", "spawn", "damaged"];
